Rename encoding -> decoder
Refs: https://bepress.atlassian.net/browse/BP-5064
