+++ decoder/url.go
@@ -5,7 +5,7 @@
 // Copyright (c) 2017 Reed O'Brien reed@reedobrien.com
 // Use of this source code is governed by an MIT-style
 // license that can be found in the LICENSE file.
-package encoding
+package decoder
 
 import (
 	"crypto/hmac"
